+++ GeosCore/cleanup.F90
@@ -50,8 +50,6 @@ SUBROUTINE CLEANUP( Input_Opt, State_Grid, ERROR, RC )
 #ifdef BPCH_DIAG
   USE CMN_O3_Mod,              ONLY : Cleanup_CMN_O3
   USE DIAG_MOD,                ONLY : CLEANUP_DIAG
-  USE DIAG51_MOD,              ONLY : CLEANUP_DIAG51
-  USE DIAG51b_MOD,             ONLY : CLEANUP_DIAG51b
 #endif
 #ifdef TOMAS
   USE TOMAS_MOD,               ONLY : CLEANUP_TOMAS  !sfarina, 1/16/13
@@ -212,8 +210,6 @@ SUBROUTINE CLEANUP( Input_Opt, State_Grid, ERROR, RC )
   ENDIF
 
 #ifdef BPCH_DIAG
-  CALL CLEANUP_DIAG()
-  CALL CLEANUP_DIAG51()
 
   CALL Cleanup_CMN_O3( RC )
   IF ( RC /= GC_SUCCESS ) THEN