+++ GeosCore/drydep_mod.F90
@@ -601,12 +601,13 @@ CONTAINS
     !$OMP END PARALLEL DO
 
     ! Set diagnostics - cnsider moving?
-    IF ( State_Diag%Archive_DryDepVel .OR. &
-         State_Diag%Archive_DryDepVelForALT1 ) THEN
+    IF ( State_Diag%Archive_DryDepVel                                   .or. &
+         State_Diag%Archive_DryDepVelForALT1                            .or. &
+         State_Diag%Archive_SatDiagnDryDepVel                         ) THEN 
 
-       !$OMP PARALLEL DO                                           &
-       !$OMP DEFAULT( SHARED                                     ) &
-       !$OMP PRIVATE( D, S, N, A, NDVZ )
+       !$OMP PARALLEL DO                                                     &
+       !$OMP DEFAULT( SHARED                                                )&
+       !$OMP PRIVATE( D, S, N, A, NDVZ                                      )
        DO D = 1, State_Chm%nDryDep
 
           ! Point to State_Chm%DryDepVel [m/s]
@@ -616,8 +617,17 @@ CONTAINS
           IF ( State_Diag%Archive_DryDepVel ) THEN
              S = State_Diag%Map_DryDepVel%id2slot(D)
              IF ( S > 0 ) THEN
-                State_Diag%DryDepVel(:,:,S) = &
-                           State_Chm%DryDepVel(:,:,NDVZ) * 100._f4
+                State_Diag%DryDepVel(:,:,S)   =                              &
+                State_Chm%DryDepVel(:,:,NDVZ) * 100.0_f4
+             ENDIF
+          ENDIF
+
+          ! Satellite diagnostic: Dry dep velocity [cm/s]
+          IF ( State_Diag%Archive_SatDiagnDryDepVel ) THEN
+             S = State_Diag%Map_SatDiagnDryDepVel%id2slot(D)
+             IF ( S > 0 ) THEN
+                State_Diag%SatDiagnDryDepVel(:,:,S)   =                      &
+                State_Chm%DryDepVel(:,:,NDVZ) * 100.0_f4
              ENDIF
           ENDIF