+++ GeosCore/gc_environment_mod.F90
@@ -437,10 +437,12 @@ CONTAINS
     USE Tagged_O3_Mod,      ONLY : Init_Tagged_O3
     USE Vdiff_Mod,          ONLY : Init_Vdiff
     USE WetScav_Mod,        ONLY : Init_WetScav
+#ifdef MODEL_CLASSIC
 #ifdef BPCH_DIAG
-    USE Diag51_Mod,         ONLY : Init_Diag51
-    USE Diag51b_Mod,        ONLY : Init_Diag51b
+    ! NOTE: Restore this call until we can remove TOMAS BPCH diagnostics
+    !  -- Bob Yantosca (03 Nov 2022)
     USE Gamap_Mod,          ONLY : Do_Gamap
+#endif
 #endif
 !
 ! !INPUT PARAMETERS:
@@ -765,6 +767,7 @@ CONTAINS
        ENDIF
     ENDIF
 
+#ifdef MODEL_CLASSIC
 #ifdef BPCH_DIAG
     !=================================================================
     ! These routines are only needed when compiling GEOS-Chem
@@ -778,15 +781,6 @@ CONTAINS
     ! Allocate and initialize variables
     CALL Ndxx_Setup( Input_Opt, State_Chm, State_Grid, RC )
 
-    ! Satellite timeseries (bpch)
-    IF ( Input_Opt%DO_ND51 ) THEN
-       CALL Init_Diag51 ( Input_Opt, State_Grid, RC )
-    ENDIF
-    IF ( Input_Opt%DO_ND51b ) THEN
-       CALL Init_Diag51b( Input_Opt, State_Grid, RC )
-    ENDIF
-
-#if defined( MODEL_CLASSIC )
     !--------------------------------------------------------------------
     ! Write out diaginfo.dat, tracerinfo.dat files for this simulation
     !